+++ database/snowflake/snowflake.go
@@ -240,7 +240,7 @@ func runesLastIndex(input []rune, target rune) int {
 	return -1
 }
 
-func (p *Snowflake) SetVersion(version int, dirty bool) error {
+func (p *Snowflake) SetVersion(version int64, dirty bool) error {
 	tx, err := p.conn.BeginTx(context.Background(), &sql.TxOptions{})
 	if err != nil {
 		return &database.Error{OrigErr: err, Err: "transaction start failed"}
@@ -276,7 +276,7 @@ func (p *Snowflake) SetVersion(version int, dirty bool) error {
 	return nil
 }
 
-func (p *Snowflake) Version() (version int, dirty bool, err error) {
+func (p *Snowflake) Version() (version int64, dirty bool, err error) {
 	query := `SELECT version, dirty FROM "` + p.config.MigrationsTable + `" LIMIT 1`
 	err = p.conn.QueryRowContext(context.Background(), query).Scan(&version, &dirty)
 	switch {
